microservices_grpc/marketplace_microservice/requirements.txt
Flask==1.1.2
grpcio==1.36.1
grpcio-tools==1.36.1
Jinja2==2.11.3
pytest==6.2.2

